LPS-115016 Split configuration install and file sync logic to 2 different classes. And move to a new configuration package.
