+++ modules/apps/static/portal-file-install/portal-file-install-impl/src/main/java/com/liferay/portal/file/install/internal/configuration/ConfigurationFileInstaller.java
@@ -12,12 +12,13 @@
  * details.
  */
 
-package com.liferay.portal.file.install.internal;
+package com.liferay.portal.file.install.internal.configuration;
 
 import com.liferay.petra.string.CharPool;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.file.install.FileInstaller;
+import com.liferay.portal.file.install.internal.DirectoryWatcher;
 import com.liferay.portal.file.install.internal.properties.TypedProperties;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
@@ -26,54 +27,30 @@ import com.liferay.portal.kernel.util.StringUtil;
 
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.io.OutputStream;
-import java.io.OutputStreamWriter;
 import java.io.Reader;
-import java.io.Writer;
 
 import java.net.URI;
 import java.net.URL;
 
-import java.util.ArrayList;
 import java.util.Dictionary;
 import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 
-import org.osgi.framework.BundleContext;
 import org.osgi.framework.Constants;
-import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.cm.Configuration;
 import org.osgi.service.cm.ConfigurationAdmin;
-import org.osgi.service.cm.ConfigurationEvent;
-import org.osgi.service.cm.ConfigurationListener;
 
 /**
  * @author Matthew Tambara
  */
-public class ConfigInstaller implements ConfigurationListener, FileInstaller {
+public class ConfigurationFileInstaller implements FileInstaller {
 
-	public ConfigInstaller(
-		BundleContext bundleContext, ConfigurationAdmin configurationAdmin,
-		FileInstallImplBundleActivator fileInstall) {
+	public ConfigurationFileInstaller(
+		ConfigurationAdmin configurationAdmin, String encoding) {
 
-		_bundleContext = bundleContext;
 		_configurationAdmin = configurationAdmin;
-		_fileInstall = fileInstall;
-
-		String encoding = _bundleContext.getProperty(
-			DirectoryWatcher.CONFIG_ENCODING);
-
-		if (encoding == null) {
-			encoding = StringPool.UTF8;
-		}
-
 		_encoding = encoding;
 	}
 
@@ -86,157 +63,6 @@ public class ConfigInstaller implements ConfigurationListener, FileInstaller {
 		}
 
 		return false;
-	}
-
-	@Override
-	public void configurationEvent(ConfigurationEvent configurationEvent) {
-		int type = configurationEvent.getType();
-
-		if (type == ConfigurationEvent.CM_UPDATED) {
-			try {
-				Configuration configuration =
-					_configurationAdmin.getConfiguration(
-						configurationEvent.getPid(), StringPool.QUESTION);
-
-				Dictionary<String, Object> dictionary =
-					configuration.getProperties();
-
-				String fileName = null;
-
-				if (dictionary != null) {
-					fileName = (String)dictionary.get(
-						DirectoryWatcher.FILENAME);
-				}
-
-				File file = null;
-
-				if (fileName != null) {
-					file = _fromConfigKey(fileName);
-				}
-
-				if ((file != null) && file.isFile()) {
-					_pidToFile.put(configuration.getPid(), fileName);
-					TypedProperties typedProperties = new TypedProperties();
-
-					try (InputStream inputStream = new FileInputStream(file);
-						Reader reader = new InputStreamReader(
-							inputStream, _encoding)) {
-
-						typedProperties.load(reader);
-					}
-
-					List<String> propertiesToRemove = new ArrayList<>();
-
-					for (String key : typedProperties.keySet()) {
-						if ((dictionary.get(key) == null) &&
-							!Objects.equals(Constants.SERVICE_PID, key) &&
-							!Objects.equals(
-								ConfigurationAdmin.SERVICE_FACTORYPID, key) &&
-							!Objects.equals(DirectoryWatcher.FILENAME, key)) {
-
-							propertiesToRemove.add(key);
-						}
-					}
-
-					Enumeration<String> enumeration = dictionary.keys();
-
-					while (enumeration.hasMoreElements()) {
-						String key = enumeration.nextElement();
-
-						if (!Objects.equals(Constants.SERVICE_PID, key) &&
-							!Objects.equals(
-								ConfigurationAdmin.SERVICE_FACTORYPID, key) &&
-							!Objects.equals(DirectoryWatcher.FILENAME, key)) {
-
-							Object value = dictionary.get(key);
-
-							typedProperties.put(key, value);
-						}
-					}
-
-					for (String key : propertiesToRemove) {
-						typedProperties.remove(key);
-					}
-
-					try (OutputStream outputStream = new FileOutputStream(file);
-						Writer writer = new OutputStreamWriter(
-							outputStream, _encoding)) {
-
-						typedProperties.save(writer);
-					}
-
-					_fileInstall.updateChecksum(file);
-				}
-			}
-			catch (Exception exception) {
-				if (_log.isWarnEnabled()) {
-					_log.warn("Unable to save configuration", exception);
-				}
-			}
-		}
-		else if (type == ConfigurationEvent.CM_DELETED) {
-			try {
-				String fileName = _pidToFile.remove(
-					configurationEvent.getPid());
-
-				File file = null;
-
-				if (fileName != null) {
-					file = _fromConfigKey(fileName);
-				}
-
-				if ((file != null) && file.isFile() && !file.delete()) {
-					throw new IOException("Unable to delete file: " + file);
-				}
-			}
-			catch (Exception exception) {
-				if (_log.isWarnEnabled()) {
-					_log.warn("Unable to delete configuration file", exception);
-				}
-			}
-		}
-	}
-
-	public void destroy() {
-		_serviceRegistration.unregister();
-	}
-
-	public void init() {
-		_serviceRegistration = _bundleContext.registerService(
-			new String[] {
-				ConfigurationListener.class.getName(),
-				FileInstaller.class.getName()
-			},
-			this, null);
-
-		try {
-			Configuration[] configurations =
-				_configurationAdmin.listConfigurations(null);
-
-			if (configurations != null) {
-				for (Configuration configuration : configurations) {
-					Dictionary<String, Object> dictionary =
-						configuration.getProperties();
-
-					String fileName = null;
-
-					if (dictionary != null) {
-						fileName = (String)dictionary.get(
-							DirectoryWatcher.FILENAME);
-					}
-
-					if (fileName != null) {
-						_pidToFile.put(configuration.getPid(), fileName);
-					}
-				}
-			}
-		}
-		catch (Exception exception) {
-			if (_log.isWarnEnabled()) {
-				_log.warn(
-					"Unable to initialize configurations list", exception);
-			}
-		}
 	}
 
 	@Override
@@ -325,10 +151,6 @@ public class ConfigInstaller implements ConfigurationListener, FileInstaller {
 		}
 
 		return null;
-	}
-
-	private File _fromConfigKey(String key) {
-		return new File(URI.create(key));
 	}
 
 	private Configuration _getConfiguration(
@@ -450,13 +272,9 @@ public class ConfigInstaller implements ConfigurationListener, FileInstaller {
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		ConfigInstaller.class);
+		ConfigurationFileInstaller.class);
 
-	private final BundleContext _bundleContext;
 	private final ConfigurationAdmin _configurationAdmin;
 	private final String _encoding;
-	private final FileInstallImplBundleActivator _fileInstall;
-	private final Map<String, String> _pidToFile = new HashMap<>();
-	private ServiceRegistration<?> _serviceRegistration;
 
 }
